Stub out preprocessor with byte copy loop for now
The preprocessor is currently experiencing bugs related to the way
Android returns video frames. It sends back frames in odd video formats
which the preprocessor currently doesn't handle right. This commit adds
some work towards fixing this.

Since it's not done yet, though, this also stubs out the normal
preprocessing implementation with a byte-for-byte copy loop that
downloads from a URI into the local file. It fulfills all the functional
requirements, so it should work okay in the event that we can't get the
real one working by the end of the term.